YDA-5165 Allow for query string in monitoring pattern portal users
The POST to /usr/gate can have, and often has, a query string. Allow for
that.

An example line from the logging in Yoda 1.8.6 (word wrapped here):

217.105.54.172 131.211.7.130 - - science.yoda.uu.nl [17/Apr/2023:13:03:05 +0200]
"POST /user/gate?redirect_target=%2Fresearch%2Fbrowse%3F HTTP/1.1" 302 856
 "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36"

The refinement of the HTTP version is probably not necessary in
any practical cases, but better shows the intent.